api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
requests==2.32.3
apscheduler==3.10.4
psutil==6.0.0
pyyaml==6.0.2
